Revert "schwab-migration v0.8.5.7 -> main"
This reverts commit bce8deecc00975a592735371517ceb9c191bf036.
