+++ account/account.go
@@ -4,11 +4,10 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/samjtro/go-trade/schwab/utils"
-	"github.com/samjtro/go-trade/schwab"
+	"github.com/samjtro/go-tda/utils"
 )
 
-// Need to migrate this whole package to schwab
+// As of 2024, these TDA endpoints are still functional. Working on migrating to Schwab before v1.0.0.
 
 type MARGIN struct {
 	Type			string
