+++ account/trade.go
@@ -1,11 +1,11 @@
-package account 
+package account
 
 import (
 	"bytes"
 	"fmt"
 	"net/http"
 
-	"github.com/samjtro/go-trade/schwab/utils"
+	"github.com/samjtro/go-tda/utils"
 )
 
 //var endpoint_replace = "https://api.tdameritrade.com/v1/accounts/%s/orders/%s"
