+++ go.mod
@@ -2,7 +2,10 @@ module github.com/samjtro/go-trade
 
 go 1.18
 
-require github.com/spf13/viper v1.12.0
+require (
+	github.com/samjtro/go-tda v0.8.3
+	github.com/spf13/viper v1.12.0
+)
 
 require (
 	github.com/fsnotify/fsnotify v1.5.4 // indirect
