+++ movers/movers.go
@@ -1,4 +1,4 @@
-package movers 
+package movers
 
 import (
 	"fmt"
@@ -7,10 +7,10 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/samjtro/go-trade/schwab/utils"
-	"github.com/samjtro/go-trade/schwab" 
+	"github.com/samjtro/go-tda/utils"
 )
 
+// As of 2024, these TDA endpoints are still functional. Working on migrating to Schwab before v1.0.0.
 type MOVER struct {
 	TICKER      string
 	DESCRIPTION string
@@ -21,7 +21,7 @@ type MOVER struct {
 }
 
 var (
-	endpoint_movers string = fmt.Sprintf(schwab.Endpoint + "/movers/%s") // Index ID
+	endpoint_movers string = "https://api.tdameritrade.com/v1/marketdata/%s/movers" // index
 )
 
 // Get returns a string; containing information on the desired index's movers per your desired direction and change type(percent or value),
