--- schwab/schwab.go
@@ -1,5 +1,0 @@
-package schwab
-
-var (
-	Endpoint string = "https://api.schwabapi.com/marketdata/v1"
-)
