+++ test/option_test.go
@@ -3,7 +3,7 @@ package test
 import (
 	"testing"
 
-	"github.com/samjtro/go-trade/option"
+	"github.com/samjtro/go-tda/option"
 )
 
 func TestOption(t *testing.T) {
